Make tests requirements reusable
Since all commands from `python setup.py` have been deprecated, it doesn't make sense to use `python setup.py test` anymore. This is the only place (AFAIK) where `tests_require` is used. As an alternative, this can be installed as an extra (as is already in the GitHub Actions workflow) with `pip install .[test]